+++ packer-images/scripts/Install-Software-V1.ps1
@@ -1,5 +1,4 @@
 choco feature enable -n allowGlobalConfirmation
-choco install azure-cli -y
 choco install microsoft-edge -y
 choco install git.install -y
 choco install netfx-4.8-devpack -y
@@ -7,5 +6,4 @@ choco install dotnet-6.0-sdk -y
 choco install powershell-core -y
 choco install winrar -y
 choco install notepadplusplus -y
-choco install sql-server-management-studio -y
 choco install azcopy10 -y