+++ packer-images/windows-2022/scripts/Execute-Regedit.ps1
@@ -1,2 +1,0 @@
-Start-Process -FilePath "regedit.exe" -ArgumentList "/s .\App-Pwsh-Context-Menu.reg" -Wait -NoNewWindow
-Start-Process -FilePath "regedit.exe" -ArgumentList "/s .\Add-Cmd-Context-Menu.reg" -Wait -NoNewWindow